+++ pkg/koyeb/services.go
@@ -7,6 +7,7 @@ import (
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/errors"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/flags_list"
+	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
@@ -213,13 +214,19 @@ func addServiceDefinitionFlags(flags *pflag.FlagSet) {
 	// Global flags
 	flags.String("type", "web", `Service type, either "web" or "worker"`)
 
-	flags.StringSlice("regions", []string{"fra"}, "Regions")
+	flags.StringSlice(
+		"regions",
+		[]string{},
+		"Add a region where the service is deployed. You can specify this flag multiple times to deploy the service in multiple regions.\n"+
+			"To update a service and remove a region, prefix the region name with '!', for example --region '!par'\n"+
+			"If the region is not specified on service creation, the service is deployed in fra\n",
+	)
 	flags.StringSlice(
 		"env",
 		[]string{},
 		"Update service environment variables using the format KEY=VALUE, for example --env FOO=bar\n"+
 			"To use the value of a secret as an environment variable, specify the secret name preceded by @, for example --env FOO=@bar\n"+
-			"To delete an environment variable, prefix its name with '!', for example --env '!FOO'",
+			"To delete an environment variable, prefix its name with '!', for example --env '!FOO'\n",
 	)
 	flags.String("instance-type", "nano", "Instance type")
 	flags.Int64("scale", 1, "Set both min-scale and max-scale")
@@ -247,7 +254,7 @@ func addServiceDefinitionFlags(flags *pflag.FlagSet) {
 		"Update service healthchecks (available for services of type \"web\" only)\n"+
 			"For HTTP healthchecks, use the format <PORT>:http:<PATH>, for example --checks 8080:http:/health\n"+
 			"For TCP healthchecks, use the format <PORT>:tcp, for example --checks 8080:tcp\n"+
-			"To delete a healthcheck, use !PORT, for example --checks '!8080'",
+			"To delete a healthcheck, use !PORT, for example --checks '!8080'\n",
 	)
 
 	// Git service
@@ -313,7 +320,6 @@ func parseServiceDefinitionFlags(flags *pflag.FlagSet, definition *koyeb.Deploym
 	definition.SetEnv(envs)
 
 	definition.SetInstanceTypes(parseInstanceType(flags, definition.GetInstanceTypes()))
-	definition.SetRegions(parseRegions(flags, definition.GetRegions()))
 
 	ports, err := parsePorts(definition.GetType(), flags, definition.Ports)
 	if err != nil {
@@ -334,6 +340,21 @@ func parseServiceDefinitionFlags(flags *pflag.FlagSet, definition *koyeb.Deploym
 		return err
 	}
 	definition.SetHealthChecks(healthchecks)
+
+	regions, err := parseRegions(flags, definition.GetRegions())
+	if err != nil {
+		return err
+	}
+	if flags.Lookup("regions").Changed && len(regions) >= 2 {
+		logrus.Warnf(
+			"Attention: you are deploying your service in %d regions (%s) which may impact your billing. If you intended to deploy your service in only one region, remove the regions you don't want to deploy to with `koyeb service update <app>/<service> --region '!<region>'`.",
+			len(regions),
+			strings.Join(regions, ", "),
+		)
+	}
+	// Scalings and environment variables refer to regions, so we must call setRegions after definition.SetScalings and definition.SetEnv.
+	setRegions(definition, regions)
+	// definition.SetRegions(parseRegions(flags, definition.GetRegions()))
 
 	err = setSource(definition, flags)
 	if err != nil {
@@ -392,19 +413,7 @@ func parseInstanceType(flags *pflag.FlagSet, currentInstanceTypes []koyeb.Deploy
 	return []koyeb.DeploymentInstanceType{*ret}
 }
 
-// Parse --regions
-func parseRegions(flags *pflag.FlagSet, currentRegions []string) []string {
-	regions, _ := flags.GetStringSlice("regions")
-	if !flags.Lookup("regions").Changed {
-		if len(currentRegions) == 0 {
-			return regions
-		}
-		return currentRegions
-	}
-	return regions
-}
-
-// parseListFlags is the generic function parsing --env, --port, --routes and --checks.
+// parseListFlags is the generic function parsing --env, --port, --routes, --checks and --regions.
 // It gets the arguments given from the command line for the given flag, then
 // builds a list of flags_list.Flag entries, and update the service
 // configuration (given in existingItems) with the new values.
@@ -526,6 +535,19 @@ func parseChecks(type_ koyeb.DeploymentDefinitionType, flags *pflag.FlagSet, cur
 		}
 	}
 	return newChecks, nil
+}
+
+// Parse --regions
+func parseRegions(flags *pflag.FlagSet, currentRegions []string) ([]string, error) {
+	newRegions, err := parseListFlags("regions", flags_list.NewRegionsListFromFlags, flags, currentRegions)
+	if err != nil {
+		return nil, err
+	}
+	// For new services, if no region is specified, add the default region
+	if !flags.Lookup("regions").Changed && len(currentRegions) == 0 {
+		newRegions = []string{"fra"}
+	}
+	return newRegions, nil
 }
 
 // Parse --min-scale and --max-scale
@@ -796,4 +818,76 @@ func parseGitSourceDockerBuilder(flags *pflag.FlagSet, builder koyeb.DockerBuild
 		builder.SetTarget(target)
 	}
 	return &builder, nil
+}
+
+// DeploymentDefinition contains the keys "env", "scalings" and "instance_types"
+// which are lists of objects with the key "scopes" containing the names of the
+// regions where the ressource should be deployed, such as:
+//
+//	"definition": {
+//		"env": [{
+//			"key": "DATABASE_URL",
+//			"scopes": ["region:fra"],
+//			"secret": "<secret value>"
+//		}],
+//		"scalings": [{
+//			"max": 1,
+//			"min": 1,
+//			"scopes": ["region:fra"]
+//		}],
+//		"instance_types": [{
+//			"scopes": ["region:fra"],
+//			"type": "nano"
+//		}],
+//	}
+//
+// setRegions updates these list of "scopes":
+// - removes the scope that are not in the list of regions (if region has been removed)
+// - add the scope that are in the list of regions but not in the list of scopes (if region has been added)
+//
+// For now, this is dumb as we do not have a feature fine grained update of what
+// is exposed for a given region. For example, while the API allows to update an
+// environment variable to have a specific value for a given region and another
+// value for another region, the CLI and the console do not allow to do that.
+func setRegions(definition *koyeb.DeploymentDefinition, regions []string) {
+	definition.SetRegions(regions)
+
+	updateScopes := func(regions []string, currentScopes []string) []string {
+		regionsMap := make(map[string]bool)
+		for _, region := range regions {
+			regionsMap[region] = false
+		}
+
+		newScopes := []string{}
+		for _, scope := range currentScopes {
+			// Append scope if it is not a region scope (even if we don't support other scopes for now)
+			if !strings.HasPrefix(scope, "region:") {
+				newScopes = append(newScopes, scope)
+			} else {
+				region := strings.TrimPrefix(scope, "region:")
+				if _, exists := regionsMap[region]; exists {
+					newScopes = append(newScopes, scope)
+					regionsMap[region] = true
+				}
+			}
+		}
+
+		// Add new regions
+		for region, exists := range regionsMap {
+			if !exists {
+				newScopes = append(newScopes, fmt.Sprintf("region:%s", region))
+			}
+		}
+		return newScopes
+	}
+
+	for idx, env := range definition.Env {
+		definition.Env[idx].Scopes = updateScopes(regions, env.Scopes)
+	}
+	for idx, scalings := range definition.Scalings {
+		definition.Scalings[idx].Scopes = updateScopes(regions, scalings.Scopes)
+	}
+	for idx, instanceTypes := range definition.InstanceTypes {
+		definition.InstanceTypes[idx].Scopes = updateScopes(regions, instanceTypes.Scopes)
+	}
 }