+++ CHANGES.md
@@ -11,6 +11,8 @@
   - https://github.com/koyeb/koyeb-cli/pull/113
 * Fix `koyeb service logs` when using the non-default organization
   - https://github.com/koyeb/koyeb-cli/issues/120
+* Fix `--regions` options of `koyeb service create` and `koyeb service update`. To deploy an existing service to a new region, use `koyeb service update <app>/<service> --region <region>`. To remove a region from a service, use `koyeb service update <app>/<service> --region '!<region>'`. On service creation, the service is still deployed to `fra` if no region is specified. If more than two regions are specified, a warning message is displayed to avoid billing surprises.
+  - https://github.com/koyeb/koyeb-cli/issues/123
 
 (v3.0.2 and before)